+++ client/src/components/footer/types.ts
@@ -1,0 +1,14 @@
+export interface ISocial {
+  link: string;
+  Icon: React.ComponentType<React.SVGProps<SVGSVGElement>>;
+}
+
+export interface INavigate {
+  link: string;
+  title: string;
+}
+
+export interface IConfigFooter {
+  social: ISocial[];
+  navigate: INavigate[];
+}
